draft: consolidate user report
