+++ app/models/user.rb
@@ -11,7 +11,7 @@ class User < ApplicationRecord
   has_many :financings, dependent: :destroy
   has_many :user_reports, dependent: :destroy
 
-  def consolidate_month_report
-    
+  def current_month_report
+    UserServices::ConsolidatedUserReport.new(self).call
   end
 end
